requirements.txt
langchain==0.3.3
langchain-openai
langchain-ollama
langchain-community
langchain-chroma
chromadb
streamlit
openai
python-dotenv
jq

